Reestructuración del proyecto: eliminados archivos antiguos y añadida nueva estructura
