--- app/modules/registro-muestra/middlewares/authMiddleware.js
@@ -1,6 +1,0 @@
-export const verificarAdmin = (req, res, next) => {
-    if (!req.usuario || req.usuario.rol !== "administrador") {
-        return res.status(403).json({ mensaje: "Acceso denegado. Se requieren permisos de administrador." });
-    }
-    next();
-};
